#9 location service WIP

+++ rac-car-service/src/main/java/pl/mg/rac/car/application/dto/command/AddCarCommand.java
@@ -2,9 +2,11 @@ package pl.mg.rac.car.application.dto.command;
 
 import pl.mg.rac.commons.value.Location;
 
+import java.math.BigDecimal;
+
 public record AddCarCommand(
         String vin,
         Location location,
-        double mileage
+        BigDecimal mileage
 ) {
 }
