+++ rac-car-service/src/main/java/pl/mg/rac/car/application/dto/command/ReturnCarCommand.java
@@ -2,9 +2,11 @@ package pl.mg.rac.car.application.dto.command;
 
 import pl.mg.rac.commons.value.Location;
 
+import java.math.BigDecimal;
+
 public record ReturnCarCommand(
         String vin,
-        double distanceTraveled,
+        BigDecimal distanceTraveled,
         String rentalId,
         Location location
 ) {
