+++ rac-car-service/src/main/java/pl/mg/rac/car/application/dto/response/ReturnCarResponse.java
@@ -2,6 +2,8 @@ package pl.mg.rac.car.application.dto.response;
 
 import pl.mg.rac.commons.value.Location;
 
-public record ReturnCarResponse(String vin, double distanceTraveled, String rentalId, Location location,
+import java.math.BigDecimal;
+
+public record ReturnCarResponse(String vin, BigDecimal distanceTraveled, String rentalId, Location location,
                                 boolean success) {
 }
