+++ rac-commons/src/main/java/pl/mg/rac/commons/event/car/payload/CarCreatedPayload.java
@@ -3,5 +3,7 @@ package pl.mg.rac.commons.event.car.payload;
 import pl.mg.rac.commons.event.RacEventPayload;
 import pl.mg.rac.commons.value.Location;
 
-public record CarCreatedPayload(String vin, Location location, double mileage) implements RacEventPayload {
+import java.math.BigDecimal;
+
+public record CarCreatedPayload(String vin, Location location, BigDecimal mileage) implements RacEventPayload {
 }
