+++ rac-commons/src/main/java/pl/mg/rac/commons/event/car/payload/CarReturnSuccessPayload.java
@@ -3,6 +3,8 @@ package pl.mg.rac.commons.event.car.payload;
 import pl.mg.rac.commons.event.RacEventPayload;
 import pl.mg.rac.commons.value.Location;
 
-public record CarReturnSuccessPayload(String vin, String rentalId, Location finalLocation, double distanceTraveled,
-                                      double mileage) implements RacEventPayload {
+import java.math.BigDecimal;
+
+public record CarReturnSuccessPayload(String vin, String rentalId, Location finalLocation, BigDecimal distanceTraveled,
+                                      BigDecimal mileage) implements RacEventPayload {
 }
